+++ test/feature/issues/issue-45-feature.js
@@ -1,61 +1,10 @@
 import { Definition } from 'bpmn-elements';
 import testHelpers from '../../helpers/testHelpers.js';
+import factory from '../../helpers/factory.js';
 
-const source = `
-  <?xml version="1.0" encoding="UTF-8"?>
-  <definitions xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL" id="sid-38422fae" targetNamespace="http://bpmn.io/bpmn">
-    <process id="Process_1" isExecutable="true">
-      <startEvent id="Event_Start">
-        <outgoing>Flow1_StartToDownstream</outgoing>
-        <outgoing>Flow2_StartToDelay</outgoing>
-      </startEvent>
-      <serviceTask id="Activity_Delay" name="delay" implementation="\${environment.services.delay}">
-        <incoming>Flow2_StartToDelay</incoming>
-        <outgoing>Flow2_DelayToUpstream</outgoing>
-      </serviceTask>
-      <parallelGateway id="Gateway_Upstream">
-        <incoming>Flow2_DelayToUpstream</incoming>
-        <outgoing>Flow2_UpstreamToDownstream</outgoing>
-      </parallelGateway>
-      <parallelGateway id="Gateway_Downstream">
-        <incoming>Flow1_StartToDownstream</incoming>
-        <incoming>Flow2_UpstreamToDownstream</incoming>
-        <outgoing>Flow_DownstreamToEnd</outgoing>
-      </parallelGateway>
-      <endEvent id="Event_End">
-        <incoming>Flow_DownstreamToEnd</incoming>
-      </endEvent>
-      <!-- branch #1 -->
-      <sequenceFlow id="Flow1_StartToDownstream" sourceRef="Event_Start" targetRef="Gateway_Downstream" />
-      <!-- branch #2 -->
-      <sequenceFlow id="Flow2_StartToDelay" sourceRef="Event_Start" targetRef="Activity_Delay" />
-      <sequenceFlow id="Flow2_DelayToUpstream" sourceRef="Activity_Delay" targetRef="Gateway_Upstream" />
-      <sequenceFlow id="Flow2_UpstreamToDownstream" sourceRef="Gateway_Upstream" targetRef="Gateway_Downstream" />
-      <!-- to end -->
-      <sequenceFlow id="Flow_DownstreamToEnd" sourceRef="Gateway_Downstream" targetRef="Event_End" />
-    </process>
-  </definitions>
-`;
+const source = factory.resource('issue-45.bpmn');
 
-const sourceWithTaskBetweenGateways = `
-  <definitions xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL" id="sid-38422fae" targetNamespace="http://bpmn.io/bpmn">
-    <process id="Process_1" isExecutable="true">
-      <startEvent id="Event_Start" />
-      <serviceTask id="Activity_Delay" name="delay" implementation="\${environment.services.delay}" />
-      <parallelGateway id="Gateway_Upstream" />
-      <serviceTask id="Activity_Between" name="between" implementation="\${environment.services.delay}" />
-      <parallelGateway id="Gateway_Downstream" />
-      <endEvent id="Event_End" />
-      <!-- branch #2 first, so the join is reached via the upstream gateway continuation after the direct branch -->
-      <sequenceFlow id="Flow2_StartToDelay" sourceRef="Event_Start" targetRef="Activity_Delay" />
-      <sequenceFlow id="Flow1_StartToDownstream" sourceRef="Event_Start" targetRef="Gateway_Downstream" />
-      <sequenceFlow id="Flow2_DelayToUpstream" sourceRef="Activity_Delay" targetRef="Gateway_Upstream" />
-      <sequenceFlow id="Flow2_UpstreamToBetween" sourceRef="Gateway_Upstream" targetRef="Activity_Between" />
-      <sequenceFlow id="Flow2_BetweenToDownstream" sourceRef="Activity_Between" targetRef="Gateway_Downstream" />
-      <sequenceFlow id="Flow_DownstreamToEnd" sourceRef="Gateway_Downstream" targetRef="Event_End" />
-    </process>
-  </definitions>
-`;
+const sourceWithTaskBetweenGateways = factory.resource('issue-45-task-between-gateways.bpmn');
 
 Feature('Issue 45 - A converging parallelGateway fires before a branch behind another parallelGateway completes', () => {
   [